+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixed
 - Coinbase parser: fixed negative values for Advanced Trade Sells.
 - MEXC parser: fees for spot trades should always be the quote asset.
+- Coinbase parser: fixed regex to ingore extra descriptions at end of trades.
 ### Added
 - Coinbase parser: added "Cash to Savings", "Savings to Cash", "Interest payout" and "Retail Simple Dust" transaction types.
 - Exodus parser: added new export format. ([#467](https://github.com/BittyTax/BittyTax/issues/467))
@@ -25,6 +26,7 @@
 - Kraken parser: added "earn" subtypes "airdrop" and "autoallocation".
 - SwissBorg parser: added "Fee Adjustment" Type.
 - Bitfinex parser: added new deposits/withdrawals export format.
+- Coinbase parser: added "Card Spend", "Credit", "Asset Migration", "Retail Eth2 Deprecation" transaction types.
 ### Changed
 - Config: fiat_income to True.
 - Price tool: CoinDesk API deprecated.